challenge-5/requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform
vertexai
pandas
pyarrow
PyPDF2
beautifulsoup4
lxml
